merge 5.5.0.RC2 development branch to master branch #309

+++ session-tutorial-init/src/session-tutorial-init-domain/src/main/java/com/example/session/domain/service/goods/GoodsService.java
@@ -48,7 +48,7 @@ public class GoodsService {
 
 		List<Goods> goodsList = Collections.emptyList();
 		if (total > 0) {
-			RowBounds rowBounds = new RowBounds(pageable.getOffset(),
+			RowBounds rowBounds = new RowBounds((int) pageable.getOffset(),
 					pageable.getPageSize());
 			goodsList = goodsRepository.findPageByCategoryId(categoryId,
 					rowBounds);
